[feat]: 홈 화면 지출 관련 usecase 적용
> 홈 최근 14일 지출 금액 합계 usecase 적용
> 홈 최근 14일 지출 금액 차트 usecase Result 타입으로 변경

+++ SolSol/PresentationLayer/Home/Presentations/Views/HomeExpenseSummaryView.swift
@@ -42,7 +42,7 @@ final class HomeExpenseSummaryView: SDView, Layoutable {
         .setDuration(3)
         .setRange(
             start: 0,
-            end: self.viewModel.totalExpense
+            end: self.viewModel.totalExpense.doubleValue
         )
         .setTextColor(color: .black100)
         .setAnimationOption(option: .linear)
